+++ src/Components/CarControl.js
@@ -92,24 +92,7 @@ export function CarControl() {
   )
 }
 
-export function imageStream(message) {
 
-  let ctx = document.getElementById("img").getContext("2d")
-  let imageData = ctx.createImageData(320, 240)
-  let messageLenght = 320 * 240
-  let ci = 0
-  for (ci = 0; ci < messageLenght; ++ci) {
-    let a = 255;
-    let r = message[3 * ci];
-    let g = message[3 * ci + 1];
-    let b = message[3 * ci + 2];
-    imageData.data[4 * ci] = r
-    imageData.data[4 * ci + 1] = g
-    imageData.data[4 * ci + 2] = b
-    imageData.data[4 * ci + 3] = a
-  }
-  ctx.putImageData(imageData, 0, 0)
-}
 
 
 export default CarControl;
